Er_mostro/llama3_rag_setup/chatbot/requirements.txt
streamlit
haystack-ai
requests
openai
sentence-transformers
python-dotenv
sqlalchemy
torch